requirements.txt
numpy==1.18.4
PyAudio==0.2.11
python-rtmidi==1.4.1
